requirements.txt
Flask
pymongo
pandas
python-dotenv
scikit-learn
flask-cors
